fix: (add) adding authentication preflow + rebuilding patient save

+++ src/app/modules/address/domain/service/AddressService.ts
@@ -1,5 +1,5 @@
-import { IAddress } from "../../../models/IAddress";
-import { AddressRepository } from "../repositories/AddressRepository";
+import { IAddress } from "../model/IAddress";
+import { AddressRepository } from "../../infra/repositories/AddressRepository";
 
 export class AddressService {
   constructor(private addressRepository: AddressRepository) {}
@@ -16,11 +16,20 @@ export class AddressService {
     return this.addressRepository.findAll();
   }
 
-  async updateAddress(addressId: number, data: IAddress): Promise<IAddress> {
-    return this.addressRepository.updateAddress(addressId, data);
+  async updateAddress(addressId: number, data: IAddress): Promise<IAddress | null> {
+    const existingAddress = await this.addressRepository.findById(addressId);
+    if (!existingAddress) {
+      throw new Error(`Address with id ${addressId} not found`);
+    }
+    const updatedAddress = { ...existingAddress, ...data };
+    return await this.addressRepository.updateAddress(addressId, updatedAddress);
   }
 
-  async deleteAddress(addressId: number): Promise<IAddress> {
-    return this.addressRepository.deleteAddress(addressId);
+  async deleteAddress(addressId: number): Promise<void> {
+    const existingAddress = await this.addressRepository.findById(addressId);
+    if (!existingAddress) {
+      throw new Error(`Address with id ${addressId} not found`);
+    }
+    await this.addressRepository.deleteAddress(addressId);
   }
 }
